% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglog.R
\name{has_pkglog}
\alias{has_pkglog}
\title{Does the folder contain a package log file?}
\usage{
has_pkglog(path = ".")
}
\arguments{
\item{path}{folder to check for a package log}
}
\description{
Does the folder contain a package log file?
}
