% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglog.R
\name{pkglog-toggle}
\alias{pkglog-toggle}
\alias{pkglog_disable}
\alias{pkglog_enable}
\title{Enable/disable automatic logging on exit}
\usage{
pkglog_disable()

pkglog_enable()
}
\description{
Enable/disable automatic logging on exit
}
\section{Functions}{
\itemize{
\item \code{pkglog_disable}: Disable logging

\item \code{pkglog_enable}: Enable logging
}}

