% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglog.R
\name{pkglog_clear}
\alias{pkglog_clear}
\title{Remove a package log file}
\usage{
pkglog_clear(path = ".")
}
\arguments{
\item{path}{folder to remove \code{pkglog.csv} from}
}
\description{
Remove a package log file
}
