% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglog.R
\name{pkglog_exit_hook}
\alias{pkglog_exit_hook}
\title{Runs at package detach or session exit}
\usage{
pkglog_exit_hook(e)
}
\arguments{
\item{e}{pkglog_env passed on exit}
}
\description{
Runs at package detach or session exit
}
