% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglog.R
\name{pkglog_write}
\alias{pkglog_write}
\title{Write the current package log to a file}
\usage{
pkglog_write(x = pkglog(), path = ".")
}
\arguments{
\item{x}{a package log \code{data.frame} to be written}

\item{path}{folder that the log file \code{pkglog.csv} is written to}
}
\description{
Write the current package log to a file
}
